Store Integers, not Floats
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@5649 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/HPRestoreItemList.java
@@ -156,7 +156,7 @@ public abstract class HPRestoreItemList
 			this.healthPerUse = healthPerUse;
 			this.purchaseCost = purchaseCost;
 
-			HPRestoreItemList.healthRestoredByName.put( restoreName, new Float( this.healthPerUse ) );
+			HPRestoreItemList.healthRestoredByName.put( restoreName, new Integer( this.healthPerUse ) );
 
 			if ( ItemDatabase.contains( restoreName ) )
 			{
